+++ CSharpGL/CSharpGL.csproj
@@ -91,10 +91,10 @@
     <None Include="GLCanvas\readme.md" />
     <Compile Include="GLAPI\GL.PrintError.cs" />
     <Compile Include="GLAPI\GL.DrawText.cs" />
+    <Compile Include="GLObjects\Framebuffer\FramebufferTarget.cs" />
     <Compile Include="GLObjects\Framebuffer\Framebuffer.Attach.Renderbuffer.cs" />
     <Compile Include="GLObjects\Framebuffer\Framebuffer.Attach.Texture.cs" />
     <Compile Include="GLObjects\Framebuffer\AttachmentLocation.cs" />
-    <Compile Include="GLObjects\Framebuffer\RenderbufferAttachment.cs" />
     <Compile Include="GLObjects\GLBuffers\IndependentBuffers\TransformFeedbackBuffer.AsVertexBuffer.cs">
       <SubType>Code</SubType>
     </Compile>
