+++ CSharpGL/GLObjects/Framebuffer/Framebuffer.Parameter.cs
@@ -5,13 +5,14 @@ namespace CSharpGL
     {
         /// <summary>
         /// Sets the size of an empty framebuffer.
+        /// These parameters only take effect when no image is attached to this framebuffer object.
         /// </summary>
         /// <param name="width"></param>
         /// <param name="height"></param>
         /// <param name="samples">how many samples?</param>
         public void SetParameter(int width, int height, int samples)
         {
-            if (glFramebufferParameteri == null) { throw new Exception(string.Format("{0} is not supported on this device!", glFramebufferParameteri)); }
+            if (glFramebufferParameteri == null) { throw new Exception(string.Format("{0} is not supported on this platform!", glFramebufferParameteri)); }
 
             //if (glFramebufferParameteri != null)
             {
