+++ CSharpGL/GLObjects/Framebuffer/Framebuffer.cs
@@ -11,14 +11,14 @@ namespace CSharpGL
         private static GLDelegates.void_int_uintN glGenFramebuffers;
         private static GLDelegates.void_uint_uint glBindFramebuffer;
         private static GLDelegates.void_uint_uint_uint_int glFramebufferTexture;
-        /// <summary>
-        /// (GLenum target​, GLenum attachment​, GLenum textarget​, GLuint texture​, GLint level);
-        /// </summary>
-        private static GLDelegates.void_uint_uint_uint_uint_int glFramebufferTexture1D;
-        /// <summary>
-        /// (GLenum target​, GLenum attachment​, GLenum textarget​, GLuint texture​, GLint level);
-        /// </summary>
-        private static GLDelegates.void_uint_uint_uint_uint_int glFramebufferTexture2D;
+        ///// <summary>
+        ///// (GLenum target​, GLenum attachment​, GLenum textarget​, GLuint texture​, GLint level);
+        ///// </summary>
+        //private static GLDelegates.void_uint_uint_uint_uint_int glFramebufferTexture1D;
+        ///// <summary>
+        ///// (GLenum target​, GLenum attachment​, GLenum textarget​, GLuint texture​, GLint level);
+        ///// </summary>
+        //private static GLDelegates.void_uint_uint_uint_uint_int glFramebufferTexture2D;
         /// <summary>
         /// (GLenum target​, GLenum attachment​, GLuint texture​, GLint level​, GLint layer);
         /// </summary>
@@ -71,8 +71,8 @@ namespace CSharpGL
             glGenFramebuffers = GL.Instance.GetDelegateFor("glGenFramebuffers", GLDelegates.typeof_void_int_uintN) as GLDelegates.void_int_uintN;
             glBindFramebuffer = GL.Instance.GetDelegateFor("glBindFramebuffer", GLDelegates.typeof_void_uint_uint) as GLDelegates.void_uint_uint;
             glFramebufferTexture = GL.Instance.GetDelegateFor("glFramebufferTexture", GLDelegates.typeof_void_uint_uint_uint_int) as GLDelegates.void_uint_uint_uint_int;
-            glFramebufferTexture1D = GL.Instance.GetDelegateFor("glFramebufferTexture1D", GLDelegates.typeof_void_uint_uint_uint_uint_int) as GLDelegates.void_uint_uint_uint_uint_int;
-            glFramebufferTexture2D = GL.Instance.GetDelegateFor("glFramebufferTexture2D", GLDelegates.typeof_void_uint_uint_uint_uint_int) as GLDelegates.void_uint_uint_uint_uint_int;
+            //glFramebufferTexture1D = GL.Instance.GetDelegateFor("glFramebufferTexture1D", GLDelegates.typeof_void_uint_uint_uint_uint_int) as GLDelegates.void_uint_uint_uint_uint_int;
+            //glFramebufferTexture2D = GL.Instance.GetDelegateFor("glFramebufferTexture2D", GLDelegates.typeof_void_uint_uint_uint_uint_int) as GLDelegates.void_uint_uint_uint_uint_int;
             glFramebufferTextureLayer = GL.Instance.GetDelegateFor("glFramebufferTextureLayer", GLDelegates.typeof_void_uint_uint_uint_int_int) as GLDelegates.void_uint_uint_uint_int_int;
             glDrawBuffers = GL.Instance.GetDelegateFor("glDrawBuffers", GLDelegates.typeof_void_int_uintN) as GLDelegates.void_int_uintN;
             glDrawBuffer = GL.Instance.GetDelegateFor("glDrawBuffer", GLDelegates.typeof_void_uint) as GLDelegates.void_uint;
@@ -82,10 +82,7 @@ namespace CSharpGL
             glCheckFramebufferStatus = GL.Instance.GetDelegateFor("glCheckFramebufferStatus", GLDelegates.typeof_uint_uint) as GLDelegates.uint_uint;
             glDeleteFramebuffers = GL.Instance.GetDelegateFor("glDeleteFramebuffers", GLDelegates.typeof_void_int_uintN) as GLDelegates.void_int_uintN;
 
-            // get max color attachment count.
-            var value = new int[1];
-            GL.Instance.GetIntegerv(GL.GL_MAX_COLOR_ATTACHMENTS, value);
-            Framebuffer.maxColorAttachmentCount = value[0];
+            Framebuffer.maxColorAttachmentCount = Framebuffer.MaxColorAttachments();
 
             Framebuffer.bindingStack.Push(null);// default framebuffer with Id = 0.
         }
@@ -146,26 +143,5 @@ namespace CSharpGL
         {
             return string.Format("Framebuffer Id: {0}", this.Id);
         }
-    }
-
-    /// <summary>
-    ///
-    /// </summary>
-    public enum FramebufferTarget : uint
-    {
-        /// <summary>
-        /// used to draw(write only) something.
-        /// </summary>
-        DrawFramebuffer = GL.GL_DRAW_FRAMEBUFFER,
-
-        /// <summary>
-        /// used to read from(read only).
-        /// </summary>
-        ReadFramebuffer = GL.GL_READ_FRAMEBUFFER,
-
-        /// <summary>
-        /// both read/write.
-        /// </summary>
-        Framebuffer = GL.GL_FRAMEBUFFER,
     }
 }
